requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.16.1  # Usa la versión CPU para ahorrar memoria
numpy==1.24.0
Pillow>=10.0.0
wikipedia>=1.4.0
fpdf
matplotlib
opencv-python
opencv-python-headless